added conversion to mebibytes + tests + refactored tests folder
