+++ scitorch/conversion/digital.py
@@ -87,7 +87,7 @@ def to_bytes(val=0.0, unit='B'):
     return ds
 
 def to_kilobytes(val=0.0, unit='KB'):
-    """Converts a value from any byte or bit format to bytes.
+    """Converts a value from any byte or bit format to kilobytes.
 
      Parameters:
         -----------
@@ -124,7 +124,7 @@ def to_kilobytes(val=0.0, unit='KB'):
 
 
 def to_megabytes(val=0.0, unit='MB'):
-    """Converts a value from any byte or bit format to bytes.
+    """Converts a value from any byte or bit format to megabytes.
 
      Parameters:
         -----------
@@ -159,7 +159,7 @@ def to_megabytes(val=0.0, unit='MB'):
     return ds
 
 def to_gigabytes(val=0.0, unit='GB'):
-    """Converts a value from any byte or bit format to bytes.
+    """Converts a value from any byte or bit format to gigabytes.
 
      Parameters:
         -----------
@@ -194,7 +194,7 @@ def to_gigabytes(val=0.0, unit='GB'):
     return ds
 
 def to_terabytes(val=0.0, unit='TB'):
-    """Converts a value from any byte or bit format to bytes.
+    """Converts a value from any byte or bit format to terabytes.
 
      Parameters:
         -----------
@@ -229,7 +229,7 @@ def to_terabytes(val=0.0, unit='TB'):
     return ds
 
 def to_petabytes(val=0.0, unit='PB'):
-    """Converts a value from any byte or bit format to bytes.
+    """Converts a value from any byte or bit format to petabytes.
 
      Parameters:
         -----------
@@ -264,7 +264,7 @@ def to_petabytes(val=0.0, unit='PB'):
     return ds
 
 def to_kibibytes(val=0.0, unit='KiB'):
-    """Converts a value from any byte or bit format to bytes.
+    """Converts a value from any byte or bit format to kibibytes.
 
      Parameters:
         -----------
@@ -280,12 +280,12 @@ def to_kibibytes(val=0.0, unit='KiB'):
         Example:
         --------
 
-        >>> gigabit = 200
-        >>> digital.to_terabytes(gigabit, 'Gbit')
-        tensor(0.0250, dtype=torch.float64)
-        >>> megabyte = 3000
-        >>> digital.to_terabytes(megabyte, 'MB')
-        tensor(0.0030, dtype=torch.float64)
+        >>> bytes = 1024
+        >>> digital.to_kibibytes(bytes, 'B')
+        tensor(1., dtype=torch.float64)
+        >>> kibibits = 1
+        >>> digital.to_kibibytes(kibibits, 'Kib')
+        tensor(0.1250, dtype=torch.float64)
 
         """
 
@@ -295,5 +295,40 @@ def to_kibibytes(val=0.0, unit='KiB'):
     else:
         ds = to_bytes(val, unit)
         ds = ds / constants.kibi
+
+    return ds
+
+def to_mebibytes(val=0.0, unit='MiB'):
+    """Converts a value from any byte or bit format to mebibytes.
+
+     Parameters:
+        -----------
+
+        val -- (int) value
+        scale -- (char) new unit
+
+        Returns:
+        --------
+
+        temp -- (Tensor) value in Kelvin scale
+
+        Example:
+        --------
+
+        >>> bytes = 1024
+        >>> digital.to_kibibytes(bytes, 'B')
+        tensor(1., dtype=torch.float64)
+        >>> kibibits = 1
+        >>> digital.to_kibibytes(kibibits, 'Kib')
+        tensor(0.1250, dtype=torch.float64)
+
+        """
+
+    # ds := digital storage
+    if unit == 'MiB':
+        ds = _create_tensor(val)
+    else:
+        ds = to_bytes(val, unit)
+        ds = ds / constants.mebi
 
     return ds
